requirements.txt
serpapi~=0.1.5
uvicorn~=0.27.1
numpy~=1.26.4
opencv-python~=4.9.0.80
pathlib~=1.0.1
fastapi~=0.110.0
pillow~=10.2.0
ultralytics~=8.1.18
tensorflow~=2.17.0